[#59] MemesServiceDto int로 되어있는 자료형 Long으로 변경

+++ backend/memetory/src/main/java/com/example/memetory/domain/memes/dto/MemesServiceDto.java
@@ -15,8 +15,8 @@ public class MemesServiceDto {
     private Long memeId;
     private String email;
     private String title;
-    private int likeCount;
-    private int commentCount;
+    private Long likeCount;
+    private Long commentCount;
 
     public Memes toEntity(Member member, Meme meme) {
         return Memes.builder()
